fix: Refactored liqudity so that combined logic is inside pages folder
