+++ app/src/components/index.ts
@@ -1,7 +1,7 @@
 export { MarketWizardCreator } from './market/sections/market_create/market_wizard_creator'
 export { MarketWizardCreatorContainer } from './market/sections/market_create/market_wizard_creator_container'
 export { MarketViewContainer } from './market/sections/market_profile/market_view_container'
-export { MarketPoolLiquidityContainer } from './market/sections/market_pooling/market_pool_liquidity_container'
+export { MarketPoolLiquidityContainer } from '../pages/market_sections/market_pool_liquidity_container'
 export { MarketBuyContainer } from './market/sections/market_buy/market_buy_container'
 export { MarketSellContainer } from './market/sections/market_sell/market_sell_container'
 export { MarketHome } from './market/sections/market_list/market_home'
