+++ app/src/components/market/sections/market_pooling/scalar_market_pool_liquidity.tsx
@@ -1,41 +1,16 @@
 import Big from 'big.js'
-import { Zero } from 'ethers/constants'
 import { BigNumber } from 'ethers/utils'
-import React, { useEffect, useMemo, useState } from 'react'
+import React, { useEffect, useState } from 'react'
 import { useHistory } from 'react-router-dom'
 import styled from 'styled-components'
 
-import { DOCUMENT_FAQ, STANDARD_DECIMALS } from '../../../../common/constants'
-import {
-  useCollateralBalance,
-  useConnectedWeb3Context,
-  useContracts,
-  useCpkAllowance,
-  useCpkProxy,
-  useFundingBalance,
-} from '../../../../hooks'
-import { getLogger } from '../../../../util/logger'
-import { getNativeAsset, pseudoNativeAssetAddress } from '../../../../util/networks'
+import { DOCUMENT_FAQ } from '../../../../common/constants'
+import { useConnectedWeb3Context } from '../../../../hooks'
+import { SharedPropsInterface } from '../../../../pages/market_sections/market_pool_liquidity_container'
+import { getNativeAsset } from '../../../../util/networks'
 import { RemoteData } from '../../../../util/remote_data'
-import {
-  bigMax,
-  bigMin,
-  calcPoolTokens,
-  calcRemoveFundingSendAmounts,
-  formatBigNumber,
-  formatNumber,
-  getInitialCollateral,
-  getUnit,
-  isDust,
-} from '../../../../util/tools'
-import {
-  AdditionalSharesType,
-  MarketDetailsTab,
-  MarketMakerData,
-  Ternary,
-  Token,
-  TransactionStep,
-} from '../../../../util/types'
+import { bigMax, bigMin, formatBigNumber, formatNumber, getUnit, isDust } from '../../../../util/tools'
+import { AdditionalSharesType, MarketDetailsTab, MarketMakerData } from '../../../../util/types'
 import { Button, ButtonContainer, ButtonTab } from '../../../button'
 import { ButtonType } from '../../../button/button_styling_types'
 import { BigNumberInput, TextfieldCustomPlaceholder } from '../../../common'
@@ -80,15 +55,57 @@ interface Props {
   switchMarketTab: (arg0: MarketDetailsTab) => void
   fetchGraphMarketMakerData: () => Promise<void>
   fetchGraphMarketUserTxData: () => Promise<void>
+  sharedProps: SharedPropsInterface
 }
 
-const logger = getLogger('Scalar Market::Fund')
-
 export const ScalarMarketPoolLiquidity = (props: Props) => {
-  const { fetchGraphMarketMakerData, fetchGraphMarketUserTxData, marketMakerData } = props
+  const { marketMakerData, sharedProps } = props
   const {
-    address: marketMakerAddress,
-    balances,
+    activeTab,
+    addFunding,
+    allowance,
+    allowanceFinished,
+    amountToFund,
+    amountToFundDisplay,
+    amountToRemove,
+    amountToRemoveDisplay,
+    collateral,
+    collateralAmountError,
+    collateralBalance,
+    depositedTokens,
+    depositedTokensTotal,
+    disableDepositButton,
+    disableDepositTab,
+    disableWithdrawButton,
+    feeFormatted,
+    fundingBalance,
+    isNegativeAmountToFund,
+    isNegativeAmountToRemove,
+    isTransactionModalOpen,
+    message,
+    poolTokens,
+    proxyIsUpToDate,
+    removeFunding,
+    setActiveTab,
+    setAmountToFund,
+    setAmountToFundDisplay,
+    setAmountToRemove,
+    setAmountToRemoveDisplay,
+    setIsTransactionModalOpen,
+    sharesAmountError,
+    sharesBalance,
+    shouldDisplayMaxButton,
+    showSetAllowance,
+    showUpgrade,
+    totalUserLiquidity,
+    txHash,
+    txState,
+    unlockCollateral,
+    upgradeFinished,
+    upgradeProxy,
+    walletBalance,
+  } = sharedProps
+  const {
     fee,
     outcomeTokenAmounts,
     outcomeTokenMarginalPrices,
@@ -101,230 +118,12 @@ export const ScalarMarketPoolLiquidity = (props: Props) => {
   } = marketMakerData
   const context = useConnectedWeb3Context()
   const history = useHistory()
-  const { account, cpk, library: provider, networkId, relay, setTxState, txHash, txState } = context
-  const { fetchBalances } = context.balances
-  const { buildMarketMaker, conditionalTokens } = useContracts(context)
-  const marketMaker = buildMarketMaker(marketMakerAddress)
+  const { networkId, relay } = context
 
-  const { proxyIsUpToDate, updateProxy } = useCpkProxy()
-  const isUpdated = RemoteData.hasData(proxyIsUpToDate) ? proxyIsUpToDate.data : true
-  const [upgradeFinished, setUpgradeFinished] = useState(false)
-
-  const resolutionDate = question.resolution.getTime()
-  const currentDate = new Date().getTime()
-  const disableDepositTab = currentDate > resolutionDate
-
-  const [activeTab, setActiveTab] = useState(disableDepositTab ? Tabs.withdraw : Tabs.deposit)
-
-  const initialCollateral = getInitialCollateral(context.networkId, marketMakerData.collateral, relay)
-  const [collateral, setCollateral] = useState<Token>(initialCollateral)
-
-  const [amountToFund, setAmountToFund] = useState<Maybe<BigNumber>>(new BigNumber(0))
-  const [amountToFundDisplay, setAmountToFundDisplay] = useState<string>('')
-  const [amountToRemove, setAmountToRemove] = useState<Maybe<BigNumber>>(new BigNumber(0))
-  const [amountToRemoveDisplay, setAmountToRemoveDisplay] = useState<string>('')
-  const [message, setMessage] = useState<string>('')
-  const [isNegativeAmountToFund, setIsNegativeAmountToFund] = useState<boolean>(false)
-  const [isNegativeAmountToRemove, setIsNegativeAmountToRemove] = useState<boolean>(false)
   const [additionalShares, setAdditionalShares] = useState<number>(0)
   const [additionalSharesType, setAdditionalSharesType] = useState<Maybe<AdditionalSharesType>>()
-  const [isTransactionModalOpen, setIsTransactionModalOpen] = useState<boolean>(false)
 
-  useEffect(() => {
-    setIsNegativeAmountToFund(formatBigNumber(amountToFund || Zero, collateral.decimals).includes('-'))
-  }, [amountToFund, collateral.decimals])
-
-  useEffect(() => {
-    setIsNegativeAmountToRemove(formatBigNumber(amountToRemove || Zero, collateral.decimals).includes('-'))
-  }, [amountToRemove, collateral.decimals])
-
-  useEffect(() => {
-    setCollateral(initialCollateral)
-    setAmountToFund(null)
-    setAmountToFundDisplay('')
-    setAmountToRemove(null)
-    setAmountToRemoveDisplay('')
-    // eslint-disable-next-line
-  }, [marketMakerData.collateral.address])
-
-  const signer = useMemo(() => provider.getSigner(), [provider])
-  const [allowanceFinished, setAllowanceFinished] = useState(false)
-  const { allowance, unlock } = useCpkAllowance(signer, collateral.address)
-
-  const { collateralBalance: maybeCollateralBalance, fetchCollateralBalance } = useCollateralBalance(
-    collateral,
-    context,
-  )
-  const collateralBalance = maybeCollateralBalance || Zero
-  const { fetchFundingBalance, fundingBalance: maybeFundingBalance } = useFundingBalance(marketMakerAddress, context)
-  const fundingBalance = maybeFundingBalance || Zero
-
-  const walletBalance = formatNumber(formatBigNumber(collateralBalance, collateral.decimals, 5), 5)
-  const sharesBalance = formatBigNumber(fundingBalance, collateral.decimals)
-
-  const hasEnoughAllowance = RemoteData.mapToTernary(allowance, allowance => allowance.gte(amountToFund || Zero))
-  const hasZeroAllowance = RemoteData.mapToTernary(allowance, allowance => allowance.isZero())
-  const showSetAllowance =
-    collateral.address !== pseudoNativeAssetAddress &&
-    !cpk?.isSafeApp &&
-    (allowanceFinished || hasZeroAllowance === Ternary.True || hasEnoughAllowance === Ternary.False)
-
-  const poolTokens = calcPoolTokens(
-    amountToFund || Zero,
-    balances.map(b => b.holdings),
-    totalPoolShares,
-  )
   const disableWithdrawTab = isDust(fundingBalance, collateral.decimals)
-
-  const sendAmountsAfterRemovingFunding = calcRemoveFundingSendAmounts(
-    amountToRemove || Zero,
-    balances.map(b => b.holdings),
-    totalPoolShares,
-  )
-
-  const depositedTokens = sendAmountsAfterRemovingFunding.length
-    ? sendAmountsAfterRemovingFunding.reduce((min: BigNumber, amount: BigNumber) => (amount.lt(min) ? amount : min))
-    : new BigNumber(0)
-  const depositedTokensTotal = depositedTokens.add(userEarnings)
-
-  const feeFormatted = useMemo(() => `${formatBigNumber(fee.mul(Math.pow(10, 2)), STANDARD_DECIMALS)}%`, [fee])
-
-  const totalUserShareAmounts = calcRemoveFundingSendAmounts(
-    fundingBalance,
-    balances.map(b => b.holdings),
-    totalPoolShares,
-  )
-
-  const totalDepositedTokens = totalUserShareAmounts.length
-    ? totalUserShareAmounts.reduce((min: BigNumber, amount: BigNumber) => (amount.lt(min) ? amount : min))
-    : new BigNumber(0)
-
-  const totalUserLiquidity = totalDepositedTokens.add(userEarnings)
-
-  const showUpgrade =
-    (!isUpdated && collateral.address === pseudoNativeAssetAddress) ||
-    (upgradeFinished && collateral.address === pseudoNativeAssetAddress)
-
-  const upgradeProxy = async () => {
-    if (!cpk) {
-      return
-    }
-
-    await updateProxy()
-    setUpgradeFinished(true)
-  }
-
-  const addFunding = async () => {
-    try {
-      if (!cpk) {
-        return
-      }
-
-      if (!account) {
-        throw new Error('Please connect to your wallet to perform this action.')
-      }
-
-      if (!cpk?.isSafeApp && collateral.address !== pseudoNativeAssetAddress && hasEnoughAllowance !== Ternary.True) {
-        throw new Error("This method shouldn't be called if 'hasEnoughAllowance' is unknown or false")
-      }
-
-      const fundsAmount = formatBigNumber(amountToFund || Zero, collateral.decimals)
-
-      setMessage(`Depositing funds: ${fundsAmount} ${collateral.symbol}...`)
-
-      setTxState(TransactionStep.waitingConfirmation)
-      setIsTransactionModalOpen(true)
-
-      await cpk.addFunding({
-        amount: amountToFund || Zero,
-        collateral,
-        marketMaker,
-      })
-
-      await fetchGraphMarketUserTxData()
-      await fetchGraphMarketMakerData()
-      await fetchFundingBalance()
-      await fetchCollateralBalance()
-      await fetchBalances()
-
-      setAmountToFund(null)
-      setAmountToFundDisplay('')
-      setMessage(`Successfully deposited ${formatNumber(fundsAmount)} ${collateral.symbol}`)
-    } catch (err) {
-      setTxState(TransactionStep.error)
-      setMessage(`Error trying to deposit funds.`)
-      logger.error(`${message} - ${err.message}`)
-    }
-  }
-
-  const removeFunding = async () => {
-    try {
-      if (!cpk) {
-        return
-      }
-
-      setTxState(TransactionStep.waitingConfirmation)
-      setIsTransactionModalOpen(true)
-      const fundsAmount = formatBigNumber(depositedTokensTotal, collateral.decimals)
-
-      setMessage(`Withdrawing funds: ${formatNumber(fundsAmount)} ${collateral.symbol}...`)
-
-      const conditionId = await marketMaker.getConditionId()
-
-      await cpk.removeFunding({
-        amountToMerge: depositedTokens,
-        conditionId,
-        conditionalTokens,
-        earnings: userEarnings,
-        marketMaker,
-        outcomesCount: balances.length,
-        sharesToBurn: amountToRemove || Zero,
-      })
-
-      await fetchGraphMarketUserTxData()
-      await fetchGraphMarketMakerData()
-      await fetchFundingBalance()
-      await fetchCollateralBalance()
-      await fetchBalances()
-
-      setAmountToRemove(null)
-      setAmountToRemoveDisplay('')
-      setMessage(`Successfully withdrew ${formatNumber(fundsAmount)} ${collateral.symbol}`)
-    } catch (err) {
-      setTxState(TransactionStep.error)
-      setMessage(`Error trying to withdraw funds.`)
-      logger.error(`${message} - ${err.message}`)
-    }
-  }
-
-  const unlockCollateral = async () => {
-    if (!cpk) {
-      return
-    }
-
-    await unlock()
-
-    setAllowanceFinished(true)
-  }
-
-  const collateralAmountError =
-    maybeCollateralBalance === null
-      ? null
-      : maybeCollateralBalance.isZero() && amountToFund?.gt(maybeCollateralBalance)
-      ? `Insufficient balance`
-      : amountToFund?.gt(maybeCollateralBalance)
-      ? `Value must be less than or equal to ${walletBalance} ${collateral.symbol}`
-      : null
-
-  const disableDepositButton =
-    !amountToFund ||
-    amountToFund?.isZero() ||
-    (!cpk?.isSafeApp && collateral.address !== pseudoNativeAssetAddress && hasEnoughAllowance !== Ternary.True) ||
-    collateralAmountError !== null ||
-    currentDate > resolutionDate ||
-    isNegativeAmountToFund
-
-  const shouldDisplayMaxButton = collateral.address !== pseudoNativeAssetAddress
 
   useEffect(() => {
     // Use floor as rounding method
@@ -363,22 +162,6 @@ export const ScalarMarketPoolLiquidity = (props: Props) => {
         : setAdditionalSharesType(AdditionalSharesType.long)
     }
   }, [collateral.decimals, outcomeTokenAmounts, amountToFund, amountToRemove, activeTab])
-
-  const sharesAmountError =
-    maybeFundingBalance === null
-      ? null
-      : maybeFundingBalance.isZero() && amountToRemove?.gt(maybeFundingBalance)
-      ? `Insufficient balance`
-      : amountToRemove?.gt(maybeFundingBalance)
-      ? `Value must be less than or equal to ${sharesBalance} pool shares`
-      : null
-
-  const disableWithdrawButton =
-    !amountToRemove ||
-    amountToRemove?.isZero() ||
-    amountToRemove?.gt(fundingBalance) ||
-    sharesAmountError !== null ||
-    isNegativeAmountToRemove
 
   return (
     <>
