+++ app/src/components/market/sections/market_profile/market_status/open.tsx
@@ -6,6 +6,7 @@ import styled from 'styled-components'
 import { useGraphMarketUserTxData } from '../../../../../hooks'
 import { WhenConnected, useConnectedWeb3Context } from '../../../../../hooks/connectedWeb3'
 import { useRealityLink } from '../../../../../hooks/useRealityLink'
+import { MarketPoolLiquidityContainer } from '../../../../../pages/market_sections/market_pool_liquidity_container'
 import { getNativeAsset, networkIds } from '../../../../../util/networks'
 import { getUnit, isDust } from '../../../../../util/tools'
 import { BalanceItem, MarketDetailsTab, MarketMakerData, OutcomeTableValue } from '../../../../../util/types'
@@ -20,7 +21,6 @@ import { MarketBondContainer } from '../../market_bond/market_bond_container'
 import { MarketBuyContainer } from '../../market_buy/market_buy_container'
 import { MarketHistoryContainer } from '../../market_history/market_history_container'
 import { MarketNavigation } from '../../market_navigation'
-import { MarketPoolLiquidityContainer } from '../../market_pooling/market_pool_liquidity_container'
 import { MarketSellContainer } from '../../market_sell/market_sell_container'
 import { MarketVerifyContainer } from '../../market_verify/market_verify_container'
